+++ builders/PathBuilder.js
@@ -17,8 +17,17 @@ export default class PathBuilder {
     configPath(_path) {
         return path.resolve(this.cwd, `config/${defineValue(_path, "")}`);
     }
+    controllersPath(_path) {
+        return path.resolve(this.cwd, `app/controllers/${defineValue(_path, "")}`);
+    }
     databasePath(_path) {
         return path.resolve(this.cwd, `database/${defineValue(_path, "")}`);
+    }
+    middlewaresPath(_path) {
+        return path.resolve(this.cwd, `app/middlewares/${defineValue(_path, "")}`);
+    }
+    modelsPath(_path) {
+        return path.resolve(this.cwd, `app/models/${defineValue(_path, "")}`);
     }
     publicPath(_path) {
         return path.resolve(this.cwd, `public/${defineValue(_path, "")}`);
@@ -34,5 +43,8 @@ export default class PathBuilder {
     }
     storagePath(_path) {
         return path.resolve(this.cwd, `storage/${defineValue(_path, "")}`);
+    }
+    validatorsPath(_path) {
+        return path.resolve(this.cwd, `app/validators/${defineValue(_path, "")}`);
     }
 }