+++ src/builders/PathBuilder.ts
@@ -24,8 +24,20 @@ export default class PathBuilder {
         return path.resolve(this.cwd, `config/${defineValue(_path, "")}`);
     }
 
+    public controllersPath(_path?: string): string {
+        return path.resolve(this.cwd, `app/controllers/${defineValue(_path, "")}`);
+    }
+
     public databasePath(_path?: string): string {
         return path.resolve(this.cwd, `database/${defineValue(_path, "")}`);
+    }
+
+    public middlewaresPath(_path?: string): string {
+        return path.resolve(this.cwd, `app/middlewares/${defineValue(_path, "")}`);
+    }
+
+    public modelsPath(_path?: string): string {
+        return path.resolve(this.cwd, `app/models/${defineValue(_path, "")}`);
     }
 
     public publicPath(_path?: string): string {
@@ -46,5 +58,9 @@ export default class PathBuilder {
 
     public storagePath(_path?: string): string {
         return path.resolve(this.cwd, `storage/${defineValue(_path, "")}`);
+    }
+
+    public validatorsPath(_path?: string): string {
+        return path.resolve(this.cwd, `app/validators/${defineValue(_path, "")}`);
     }
 }